# docs/swagger.yaml
basePath: /api
definitions:
  SangXanh_pkg_dto.Address:
    properties:
      address_json:
        type: string
      is_default_address:
        type: boolean
      name:
        type: string
      phone:
        type: string
    type: object
  SangXanh_pkg_dto.AuthResponse:
    properties:
      access_token:
        type: string
      refresh_token:
        type: string
    type: object
  SangXanh_pkg_dto.CartCreateRequest:
    properties:
      product_option_id:
        type: string
      quantity:
        type: integer
      user_id:
        type: string
    type: object
  SangXanh_pkg_dto.CartResponse:
    properties:
      created_at:
        type: string
      deleted_at:
        type: string
      id:
        type: string
      product_option:
        $ref: '#/definitions/SangXanh_pkg_dto.ProductOption'
      product_option_id:
        type: string
      quantity:
        type: integer
      updated_at:
        type: string
      user_id:
        type: string
    type: object
  SangXanh_pkg_dto.CartUpdate:
    properties:
      id:
        type: string
      quantity:
        type: integer
    type: object
...
  SangXanh_pkg_dto.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
      username:
        type: string
    type: object
...
  SangXanh_pkg_dto.ProductCreated:
    properties:
      category_id:
        type: string
      content:
        type: string
      description:
        type: string
      discount:
        type: number
      discount_type:
        type: string
      image_detail:
        type: string
      metadata:
        additionalProperties:
          type: string
        type: object
      name:
        type: string
      price:
        type: number
      product_code:
        type: string
      questions:
        items:
          additionalProperties:
            type: string
          type: object
        type: array
      thumbnail:
        type: string
    type: object
  SangXanh_pkg_dto.ProductOption:
    properties:
      created_at:
        type: string
      detail:
        items:
          $ref: '#/definitions/SangXanh_pkg_dto.ProductOptionDetail'
        type: array
      id:
        type: string
      metadata:
        items:
          additionalProperties:
            type: string
          type: object
        type: array
      name:
        type: string
      price:
        type: number
      product_id:
        type: string
      updated_at:
        type: string
    type: object
...
  SangXanh_pkg_dto.ProductOptionDetail:
    properties:
      name:
        type: string
      variant_id:
        type: string
    type: object
...
  SangXanh_pkg_dto.ProductResponse:
    properties:
      category_id:
        type: string
      content:
        type: string
      created_at:
        type: string
      description:
        type: string
      discount:
        type: number
      discount_type:
        type: string
      id:
        type: string
      image_detail:
        type: string
      name:
        type: string
      price:
        type: number
      product_code:
        type: string
      question:
        items:
          additionalProperties:
            type: string
          type: object
        type: array
      thumbnail:
        type: string
      updated_at:
        type: string
    type: object
...
  SangXanh_pkg_dto.ProductUpdated:
    properties:
      category_id:
        type: string
      content:
        type: string
      description:
        type: string
      discount:
        type: number
      discount_type:
        type: string
      id:
        type: string
      image_detail:
        type: string
      metadata:
        additionalProperties:
          type: string
        type: object
      name:
        type: string
      price:
        type: number
      product_code:
        type: string
      questions:
        items:
          additionalProperties:
            type: string
          type: object
        type: array
      thumbnail:
        type: string
    type: object
...
  SangXanh_pkg_dto.RefreshTokenRequest:
    properties:
      refresh_token:
        type: string
    type: object
...
  SangXanh_pkg_dto.UserInfo:
    properties:
      address:
        items:
          $ref: '#/definitions/SangXanh_pkg_dto.Address'
        type: array
      avatar:
        type: string
      basic_address:
        type: string
      email:
        type: string
      full_name:
        type: string
      id:
        type: string
      phone:
        type: string
      role:
        type: string
      status:
        $ref: '#/definitions/SangXanh_pkg_enum.Status'
      username:
        type: string
    type: object
...
  SangXanh_pkg_enum.Status:
    enum:
    - active
    - inactive
    type: string
    x-enum-varnames:
    - Active
    - Inactive
host: localhost:8080
info:
  contact:
    email: support@sangxanh.com
    name: API Support
    url: http://www.sangxanh.com/support
  description: This is the SangXanh API server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: SangXanh API
  version: "1.0"
paths:
  /auth/current-user:
    get:
      consumes:
      - application/json
      description: Retrieve the authenticated user's information
      produces:
      - application/json
      responses:
        "200":
          description: Current user information
          schema:
            $ref: '#/definitions/SangXanh_pkg_dto.UserInfo'
        "401":
          description: Unauthorized - invalid or missing token
          schema:
            additionalProperties: true
            type: object
        "429":
          description: Too many requests
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get current user information
      tags:
      - Authentication
  /auth/login:
    post:
      consumes:
      - application/json
      description: Authenticate user with username/email and password
      parameters:
      - description: Login credentials
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/SangXanh_pkg_dto.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Successfully authenticated
          schema:
            $ref: '#/definitions/SangXanh_pkg_dto.AuthResponse'
        "400":
          description: Invalid request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Invalid credentials
          schema:
            additionalProperties: true
            type: object
        "429":
          description: Too many requests
          schema:
            additionalProperties: true
            type: object
      summary: User login
      tags:
      - Authentication
  /auth/refresh:
    post:
      consumes:
      - application/json
      description: Generate a new access token using a refresh token
      parameters:
      - description: Refresh token
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/SangXanh_pkg_dto.RefreshTokenRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Successfully refreshed token
          schema:
            $ref: '#/definitions/SangXanh_pkg_dto.AuthResponse'
        "400":
          description: Invalid request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Invalid or expired refresh token
          schema:
            additionalProperties: true
            type: object
        "429":
          description: Too many requests
          schema:
            additionalProperties: true
            type: object
      summary: Refresh access token
      tags:
      - Authentication
  /cart:
    get:
      consumes:
      - application/json
      description: Get all cart items for the authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: List of cart items
          schema:
            items:
              $ref: '#/definitions/SangXanh_pkg_dto.CartResponse'
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: List user's cart items
      tags:
      - Cart
  /cart/create:
    post:
      consumes:
      - application/json
      description: Add a product option to the user's cart
      parameters:
      - description: Cart item data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/SangXanh_pkg_dto.CartCreateRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created cart item
          schema:
            $ref: '#/definitions/SangXanh_pkg_dto.CartResponse'
        "400":
          description: Invalid request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Add item to cart
      tags:
      - Cart
  /cart/delete:
    delete:
      consumes:
      - application/json
      description: Delete a cart item by ID
      parameters:
      - description: Cart item ID
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Cart item deleted successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Cart item not found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Remove item from cart
      tags:
      - Cart
  /cart/update:
    put:
      consumes:
      - application/json
      description: Update the quantity of a cart item
      parameters:
      - description: Cart update data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/SangXanh_pkg_dto.CartUpdate'
      produces:
      - application/json
      responses:
        "200":
          description: Updated cart item
          schema:
            $ref: '#/definitions/SangXanh_pkg_dto.CartResponse'
        "400":
          description: Invalid request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Cart item not found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Update cart item quantity
      tags:
      - Cart
...
  /product:
    get:
      consumes:
      - application/json
      description: Get a paginated list of products with optional filters
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Page size
        in: query
        name: page_size
        type: integer
      - description: Filter by product name
        in: query
        name: name
        type: string
      - description: Filter by category ID
        in: query
        name: category_id
        type: string
      - description: Minimum price
        in: query
        name: min_price
        type: number
      - description: Maximum price
        in: query
        name: max_price
        type: number
      produces:
      - application/json
      responses:
        "200":
          description: Paginated list of products
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request
          schema:
            additionalProperties: true
            type: object
      summary: List all products
      tags:
      - Products
...
  /product/{id}:
    get:
      consumes:
      - application/json
      description: Retrieve a specific product by its ID
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Product details
          schema:
            $ref: '#/definitions/SangXanh_pkg_dto.ProductResponse'
        "404":
          description: Product not found
          schema:
            additionalProperties: true
            type: object
      summary: Get product by ID
      tags:
      - Products
  /product/create:
    post:
      consumes:
      - application/json
      description: Create a new product (admin only)
      parameters:
      - description: Product data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/SangXanh_pkg_dto.ProductCreated'
      produces:
      - application/json
      responses:
        "201":
          description: Created product
          schema:
            $ref: '#/definitions/SangXanh_pkg_dto.ProductResponse'
        "400":
          description: Invalid request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden - admin only
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Create a new product
      tags:
      - Products
  /product/delete:
    delete:
      consumes:
      - application/json
      description: Delete a product by ID (admin only)
      parameters:
      - description: Product ID
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Product deleted successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden - admin only
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Product not found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Delete a product
      tags:
      - Products
  /product/update:
    put:
      consumes:
      - application/json
      description: Update an existing product (admin only)
      parameters:
      - description: Product update data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/SangXanh_pkg_dto.ProductUpdated'
      produces:
      - application/json
      responses:
        "200":
          description: Updated product
          schema:
            $ref: '#/definitions/SangXanh_pkg_dto.ProductResponse'
        "400":
          description: Invalid request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden - admin only
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Product not found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Update a product
      tags:
      - Products
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"